+++ packages/vscode/golang/CHANGELOG.md
@@ -1,11 +1,11 @@
 # Change Log
 
-## [1.0.0] - 2025-07-22
+## [1.0.0] - 2025-07-23
 
 ### Added
 - Initial release of golang extension pack for vscode
 - 2 required extensions
-- 4 optional extensions
+- 7 optional extensions
 
 ### Extensions Included
 
@@ -18,3 +18,6 @@
 - [Go Test Explorer](https://marketplace.visualstudio.com/items?itemName&#x3D;premparihar.gotestexplorer) - Go Test Explorer for Visual Studio Code
 - [Go Fill Struct](https://marketplace.visualstudio.com/items?itemName&#x3D;davidbarratt.go-fill-struct) - Fill struct literals with default values
 - [Go Template Syntax Highlighter](https://marketplace.visualstudio.com/items?itemName&#x3D;karyan40024.gotmpl-syntax-highlighter) - Go Template Syntax Highlighter for Reply
+- [Go Doc](https://marketplace.visualstudio.com/items?itemName&#x3D;msyrus.go-doc) - Show documentation of go symbols and packages
+- [Go Critic](https://marketplace.visualstudio.com/items?itemName&#x3D;neverik.go-critic) - Integration for the go-critic golang linter
+- [Go Asm](https://marketplace.visualstudio.com/items?itemName&#x3D;quillaja.goasm) - Syntax highlighting &amp; autocomplete for Go assembly